Move KmTypes to internal package
